new component AddFileCard with delegated file handling
